[Sources] Implement message reading for file data source
The data source reads the contents of the provided file in string
chunks. After reading a chunk, it splits the text into multiple JSON
string messages which are appended to an internal queue. When the data
source is queried for a message the first element of the queue is popped
and deserialized.
Any additional string data that does not fully form a JSON message is
written to a string stream, whose value is prepended to the next text
chunk.